requirements.txt
streamlit==1.37.1
pandas==2.2.2
scikit-learn==1.5.1
langdetect==1.0.9
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5
google-generativeai==0.6.0
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==4.25.3
google-api-python-client==2.130.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0